Add support for ~/.cfg config repo
This is a much more elegant way to manage dotfiles in ~/ without having
to use a bunch of symlinks, by ensuring that the repo information is
stored in .cfg instead of in .git. It uses a couple of aliases to make
this easy, allowing interaction with dotfiles repo via "config" command
instead of git.

See https://www.atlassian.com/git/tutorials/dotfiles for more
information.