Version-gate lsp config
Neovim LSP support was added in version 0.5, however older Ubuntu/Debian
ships with neovim 0.4.4, in which case it generates an error you have to
go past every time you start to edit a file.

Version-gating that chunk of config to neovim 0.5 fixes the problem, and
allows us to use LSPs when running a new enough version of neovim.