Add new neovim config for coding
I've checked it for basic functionality on:
- rust
- golang
- typescript (incl awscdk)

This includes showing signatures while writing the function call.

This is not going to be found by a default nvim, nvim application name
has to be changed to nvim-kickstart for now (to avoid utterly my
existing mess of a config)

This + i3 should be runnable in a linux VM on my macbook. I hope.